+++ src/client/components/validity_panel/validity_panel.js
@@ -70,7 +70,7 @@ Template.validity_panel.helpers({
     parmsValidities( p ){
         return {
             ...this,
-            classes: 'btn btn-sm btn-outline-secondary',
+            classes: 'btn btn-sm btn-outline-primary',
             period: p
         };
     }